created InvoiceDetails component
